refactor: make copy-pasted code common
